**Context**
- My wife has been asking for an app that will let her keep track of what she likes at restaurants we go to so she doesn't forget and order something she knows she doesn't like.
- Since I am using Grok for coding this month I fed it this prompt (yep, typos and all):
Can you generate my a web page that uses local storage (cookies or something) that you can make notes on what you like to eat in certain restaurants.  So basically there should be a + button that lets you add a new restaurant.  Then there is a field that let's you type notes for that restaurant.  It should be easy to add new restaurants, notes, and to navigate the restaurants you have already added to see the notes.  It should not need a server and work  on github pages.
- It came up with a rudimentary solution that may just get the job done but definitelly isn't pretty.
- We will see what my wife says!

**Objective**

- Write a Medium blog

**Details**

- topic: Grok makes a restaurant notes app
- tone: friendly, very humorous, light and conversational
- author: AI Persona Dave LumAI
- author should not be included in the blog content
- don't include any other data before or after the blog
- Don't repeat yourself.  If you have made a point once, don't bring it up a second time unless it is super important and then you better clarify why you are bringing it up again.
- Guide people to click the "Follow" link at the top of the page.
- include a link for the page Grok generated: https://lumaiere.com/RestaurantNotes/

**Questions**:

- How many iterations or additional prompts would it take to improve this website?
- Did this save me time?
- Did it make my wife happy?