Context
Almost two weeks exactly and we are in the midst of another power outage from another hurricane.
Hurricane Milton is about 200 miles south west of us moving at 16 miles per hour in our direction and speeding up.

Objective

Write a Medium blog

Details

topic: The history of hurricanes hitting tampa bay
tone: friendly, humorous, light and conversational
author: AI Persona Dave LumAI
author should NOT be included in the blog content (no intros or mention of Dave LumAI)
don't include any other data before or after the blog
Don't repeat yourself.  If you have made a point once, don't bring it up a second time unless it is super important and then you better clarify why you are bringing it up again.
Tell people in the path of the hurricane to be safe.
provide some references to huuricane preparation
provide some references to what to do if stuck in a hurricane with no power and no more preparation possible.
Guide people to comment

Questions

When is the last hurricane that hit tampa bay directly?
How many hurricanes in last 100 years were supposed to hit tampa bay and changed course?
When is the earliest season hurricane to hit Tampa Bay?
When is the latest season hurricane to hit tampa bay?
Which was the most destructive hurricane to hit tampa bay in terms of:
loss of life?
loss of property?
Are there any other interesting facts about Hurricanes in Tampa Bay?